Implemented i2c_read
The Master can now send slaves an id asking for a specific package of
data.

+++ com-module/i2c_slave.c
@@ -44,7 +44,7 @@ void clear_twint( void )
 
 /* Read data from bus. */
 void read ( void ) {
-	uint8_t recv_data=TWDR;	// Load incoming data into recv_data.
+        uint16_t recv_data=TWDR;	// Load incoming data into recv_data.
 	
 	/* Select buffer */
 	if ( buffer ) {
